Refatoração Admin - Limite de 6 para Cod Setor, Departamento

+++ database/migrations/2012_11_30_084800_cria_tabela_setores.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->bigIncrements('id');
             $table->unsignedBigInteger('departamento');
 
-            $table->integer('cod_setor')->unique();
+            $table->bigInteger('cod_setor')->unique();
             $table->string('nome')->unique();
             $table->boolean('ativo')->default(true);
